requirements.txt
requests
python-telegram-bot
emoji
jira
